Update header file and memory functions.
Handle NULL pointers in ft_bzero.c ft_calloc.c and ft_memset.c
Update header file with more functions and their documentation.

+++ memory/ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: ribana-b <ribana-b@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 11:17:59 by ribana-b          #+#    #+#             */
-/*   Updated: 2023/10/04 11:34:10 by ribana-b         ###   ########.fr       */
+/*   Updated: 2023/11/15 14:59:35 by ribana-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/bfl.h"
 
-void	*ft_calloc(unsigned int bytes, unsigned int sizetype)
+void	*ft_calloc(size_t bytes, size_t sizetype)
 {
 	void	*ptr;
